requirements.txt
numpy
langchain
streamlit
langchain_community
sentence-transformers
weaviate-client
llama-cpp-python
pathlib